requirements.txt
loguru
pywebview
